added startTab.jsx for Start Tab
