+++ frontend/App.jsx
@@ -4,9 +4,8 @@ import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { MaterialCommunityIcons, Ionicons, FontAwesome5 } from '@expo/vector-icons';
-import {
-  HomeTab, Start,
-} from './homeTabs';
+import HomeTab from './homeTabs';
+import Start from './startTab';
 import DiagnosticsTab from './diagnosticsTab';
 import SettingsTab from './settingsTab';
 import DataTab from './dataTab';
