+++ frontend/homeTabs.jsx
@@ -3,35 +3,14 @@
 // are navigated to (except for the Settings Screen)
 
 import * as React from 'react';
-import { LinearGradient } from 'expo-linear-gradient';
 import { Feather } from '@expo/vector-icons';
 import { sampleData } from './sampleData.json';
 import {
-  CardTitle, CardText, HomeCardData, HomeCardData2, HomeCardTitle, HomeCardView,
-  PageTitle, PageView, StartView,
+  CardText, HomeCardData, HomeCardData2, HomeCardTitle, HomeCardView,
+  PageTitle, PageView,
 } from './Themes';
 
-export function Start() {
-  const [time, setTime] = React.useState(new Date().toTimeString().split(' ')[0]);
-
-  React.useEffect(() => {
-    const timer = setInterval(() => {
-      setTime(new Date().toTimeString().split(' ')[0]);
-    }, 100);
-    return () => clearInterval(timer);
-  });
-
-  return (
-    <StartView>
-      <LinearGradient colors={['#F9F6FF', '#CFDFF7']} style={{ flex: 1, padding: 20, paddingTop: 80 }}>
-        <CardTitle center>Begin Sleep Session</CardTitle>
-        <PageTitle center>{time}</PageTitle>
-      </LinearGradient>
-    </StartView>
-  );
-}
-
-export function HomeTab({ navigation }) {
+export default function HomeTab({ navigation }) {
   const latestData = sampleData[sampleData.length - 1];
   return (
     <PageView center>
